# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WebAuthn/Passkey Authentication API with HttpOnly JWT cookies",
        "title": "Go Auth Core API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/me": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Restituisce le informazioni dell'utente autenticato. Richiede un cookie access_token valido.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "Informazioni utente",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Non autenticato o token scaduto",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/login/begin": {
            "post": {
                "description": "Initiates the WebAuthn authentication flow. Returns options to be passed to navigator.credentials.get() in the browser.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Start Passkey Login",
                "parameters": [
                    {
                        "description": "User Email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PublicKeyCredentialRequestOptions for WebAuthn",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid email",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login/finish": {
            "post": {
                "description": "Completes the WebAuthn authentication flow. If login is successful, sets two HttpOnly cookies: access_token (15 min) and refresh_token (7 days).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Complete Passkey Login",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user@example.com",
                        "description": "User Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "WebAuthn response from navigator.credentials.get()",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login completed, JWT cookies set",
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginSuccessResponse"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "refresh_token=\u003cJWT\u003e; Path=/auth; HttpOnly; SameSite=Lax"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing email",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Token generation error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "description": "Uses the refresh token to obtain a new access token. The refresh token is rotated for security.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Renew Access Token",
                "responses": {
                    "200": {
                        "description": "Token renewed",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponse"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "refresh_token=\u003cJWT\u003e; Path=/auth; HttpOnly; SameSite=Lax"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid, or expired refresh token",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register/begin": {
            "post": {
                "description": "Initiates the WebAuthn registration flow. Returns options to be passed to navigator.credentials.create() in the browser.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Start Passkey Registration",
                "parameters": [
                    {
                        "description": "User Email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PublicKeyCredentialCreationOptions for WebAuthn",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing email",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register/finish": {
            "post": {
                "description": "Completes the WebAuthn registration flow. The body must contain the raw response from navigator.credentials.create().",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Complete Passkey Registration",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user@example.com",
                        "description": "User Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "WebAuthn response from navigator.credentials.create()",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration completed",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Missing email",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Registration failed",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register/verify-otp": {
            "post": {
                "description": "Verifies the OTP sent to email and returns registration options.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify OTP and Continue Registration",
                "parameters": [
                    {
                        "description": "Email and OTP",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PublicKeyCredentialCreationOptions",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.EmailRequest": {
            "description": "Request body with user email",
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "User's email address\nrequired: true\nexample: user@example.com",
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "dtos.ErrorResponse": {
            "description": "Error details",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Optional error code (e.g., TOKEN_EXPIRED)",
                    "type": "string",
                    "example": "TOKEN_EXPIRED"
                },
                "error": {
                    "description": "Error message\nexample: Invalid request format",
                    "type": "string",
                    "example": "Invalid request format"
                }
            }
        },
        "dtos.LoginSuccessResponse": {
            "description": "Response after successful authentication",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message\nexample: Login successful! 🔐",
                    "type": "string",
                    "example": "Login successful! 🔐"
                },
                "user": {
                    "description": "User information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.UserInfo"
                        }
                    ]
                }
            }
        },
        "dtos.MessageResponse": {
            "description": "Generic response message",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Response message\nexample: Operation successful",
                    "type": "string",
                    "example": "Operation successful"
                }
            }
        },
        "dtos.UserInfo": {
            "description": "Authenticated user details",
            "type": "object",
            "properties": {
                "email": {
                    "description": "User email\nexample: user@example.com",
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "description": "Unique user ID\nexample: 1",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dtos.UserResponse": {
            "description": "Response containing user info",
            "type": "object",
            "properties": {
                "user": {
                    "description": "Authenticated user info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.UserInfo"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT token in HttpOnly cookie (set automatically on login)",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}